requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlmodel==0.0.14
pytest==7.4.4
python-dotenv==1.0.0
httpx==0.26.0